+++ wikitextparser/__init__.py
@@ -1,7 +1,7 @@
 """Initialize the wikitextparser."""
 
 # Scheme: [N!]N(.N)*[{a|b|rc}N][.postN][.devN]
-__version__ = '0.23.1.dev0'
+__version__ = '0.24.0'
 
 from ._parameter import Parameter
 from ._argument import Argument